+++ lib/helpers/show_date_picker.dart
@@ -13,6 +13,7 @@ void showMaterialDatePicker({
   String okButtonLabel,
   String cancelButtonLabel,
   ValueChanged<DateTime> onChanged,
+  VoidCallback onCancelled,
 }) {
   if (firstDate == null) firstDate = DateTime(1990, 1, 1);
   if (lastDate == null) lastDate = DateTime(2050, 12, 31);
@@ -26,5 +27,6 @@ void showMaterialDatePicker({
     confirmText: okButtonLabel,
   ).then((selection) {
     if (onChanged != null && selection != null) onChanged(selection);
+    if (onCancelled != null && selection == null) onCancelled();
   });
 }