+++ lib/helpers/show_responsive_dialog.dart
@@ -1,6 +1,7 @@
 // Copyright (c) 2018, codegrue. All rights reserved. Use of this source code
 // is governed by the MIT license that can be found in the LICENSE file.
 
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_material_pickers/dialogs/responsive_dialog.dart';
 
@@ -19,6 +20,7 @@ void showMaterialResponsiveDialog({
   double maxShortSide,
   bool hideButtons = false,
   ValueChanged<List<String>> onChanged,
+  VoidCallback onCancelled,
 }) {
   showDialog<List<String>>(
     context: context,
@@ -41,5 +43,6 @@ void showMaterialResponsiveDialog({
     },
   ).then((selection) {
     if (onChanged != null && selection != null) onChanged(selection);
+    if (onCancelled != null && selection == null) onCancelled();
   });
 }