+++ lib/helpers/show_swatch_picker.dart
@@ -19,6 +19,7 @@ void showMaterialSwatchPicker({
   double maxLongSide,
   double maxShortSide,
   ValueChanged<Color> onChanged,
+  VoidCallback onCancelled,
 }) {
   showDialog<Color>(
     context: context,
@@ -47,5 +48,6 @@ void showMaterialSwatchPicker({
     },
   ).then((selection) {
     if (onChanged != null && selection != null) onChanged(selection);
+    if (onCancelled != null && selection == null) onCancelled();
   });
 }